fix: remove include bits/types/FILE.h from grammar.h and rename algo_registrator
