+++ src/cfpq_algorithms/algo_registrator.h
